Remove previous implementation
